requirements.txt
Faker==37.4.0
kafka-python==2.2.11
tzdata==2025.2
